setup flask in main; add handlers
